# docs/swagger.yaml
definitions:
...
  request.ParameterEdit:
    properties:
      confirmation:
        description: 交易确认数
        example: 12
        type: integer
      create_at:
        description: 更新时间
        type: string
      gas_limit:
        description: gas Limit
        example: 1
        type: integer
      gas_price:
        description: GasPrice
        example: 88
        type: number
    type: object
  request.PickAdd:
    properties:
      amount:
        description: 数量
        example: 1
        type: number
      pick_address:
        description: 收款地址
        example: 0x1243cfsadfcsd
        type: string
      remark:
        description: 备注
        example: 备注
        type: string
      type:
        description: 类型 1：代发 0：提领
        example: 1
        type: integer
      user_id:
        description: 用户id
        example: 1
        type: integer
    required:
    - amount
    - pick_address
    - remark
    - type
    - user_id
    type: object
...
  request.SendAdd:
    properties:
      amount:
        description: 数量
        example: 1
        type: number
      order_code:
        description: 商户订单编号
        example: asfasgdsasfgas
        type: string
      pick_address:
        description: 收款地址
        example: 0x1243cfsadfcsd
        type: string
      remark:
        description: 备注
        example: 备注
        type: string
      type:
        description: 类型 1：代发 0：提领
        example: 1
        type: integer
      user_id:
        description: 用户id
        example: 1
        type: integer
    required:
    - amount
    - order_code
    - pick_address
    - remark
    - type
    - user_id
    type: object
  request.SendEdit:
    properties:
      status:
        description: 状态 0：等待中，1：执行中，2：成功，3：已取消，4：失败
        example: 3
        type: integer
    required:
    - status
    type: object
...
  response.MerchantPickList:
    properties:
      amount:
        description: 金额
        type: number
      create_at:
        description: 建立时间
        type: string
      del_money:
        description: 扣除商户余额
        type: number
      fee:
        description: Gas手续费
        type: number
      finish_time:
        description: 完成时间
        type: string
      id:
        description: ID
        type: integer
      id_code:
        description: 系统编号
        type: string
      order_code:
        description: 商户订单编号
        type: string
      real_name:
        description: 组织名称
        type: string
      remark:
        description: 备注
        type: string
      send_address:
        description: 代发地址
        type: string
      status:
        description: 状态 0：等待中，1：执行中，2：成功，3：已取消，4：失败
        type: integer
      txid:
        description: 交易hash
        type: string
      type:
        description: 类型 1：提领，2：代发
        type: integer
      user_id:
        description: 商户id
        type: integer
    type: object
  response.MerchantPickSendPage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        items:
          $ref: '#/definitions/response.MerchantPickList'
        type: array
      per_page:
        description: 当前页码
        type: integer
      total:
        description: 总共多少页
        type: integer
      total_amount:
        description: 提领总额
        type: number
      total_fee:
        description: 总手续费
        type: number
      total_reduce:
        description: 总减少金额
        type: number
    type: object
...
  response.PickList:
    properties:
      amount:
        description: 金额
        type: number
      create_at:
        description: 建立时间
        type: string
      finish_time:
        description: 完成时间
        type: string
      id:
        description: ID
        type: integer
      id_code:
        description: 系统编号
        type: string
      real_name:
        description: 组织名称
        type: string
      send_address:
        description: 代发地址
        type: string
      status:
        description: 状态 0：等待中，1：执行中，2：成功，3：已取消，4：失败
        type: integer
      txid:
        description: 交易hash
        type: string
      type:
        description: 类型 1：提领，2：代发
        type: integer
      user_id:
        description: 商户id
        type: integer
    type: object
  response.PickUpPage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        items:
          $ref: '#/definitions/response.PickList'
        type: array
      per_page:
        description: 当前页码
        type: integer
      total:
        description: 总共多少页
        type: integer
    type: object
...
  response.SendList:
    properties:
      amount:
        description: 金额
        type: number
      create_at:
        description: 建立时间
        type: string
      del_money:
        description: 扣除商户余额
        type: number
      fee:
        description: 手续费
        type: number
      finish_time:
        description: 完成时间
        type: string
      id:
        description: ID
        type: integer
      id_code:
        description: 系统编号
        type: string
      order_code:
        description: 商户订单编号
        type: string
      real_name:
        description: 组织名称
        type: string
      remark:
        description: 备注
        type: string
      send_address:
        description: 代发地址
        type: string
      status:
        description: 状态 0：等待中，1：执行中，2：成功，3：已取消，4：失败
        type: integer
      txid:
        description: 交易hash
        type: string
      type:
        description: 类型 1：提领，2：代发
        type: integer
      user_id:
        description: 商户id
        type: integer
    type: object
  response.SendPage:
    properties:
      current_page:
        description: 每页显示多少条
        type: integer
      data:
        items:
          $ref: '#/definitions/response.SendList'
        type: array
      per_page:
        description: 当前页码
        type: integer
      total:
        description: 总共多少页
        type: integer
    type: object
...
info:
  contact: {}
  license: {}
  title: y2pay
paths:
...
  /merchantPick:
    get:
      parameters:
      - description: -1：等待中，1:执行中，2：成功，3：取消，4，失败
        in: query
        name: status
        type: integer
      - description: 0：所有，1：代发，2：提领
        in: query
        name: type
        type: integer
      - description: 商户订单编号
        in: query
        name: orderCode
        type: string
      - description: 商户id
        in: query
        name: userId
        required: true
        type: integer
      - description: 起
        in: query
        name: from_date
        type: string
      - description: 至
        in: query
        name: to_date
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MerchantPickSendPage'
      summary: 商户提领、代发列表
      tags:
      - 商户提领代发管理
    post:
      parameters:
      - description: 商户提领
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.PickAdd'
      produces:
      - application/json
      responses:
        "200": {}
      summary: 提领
      tags:
      - 商户提领代发管理
  /merchantPick/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MerchantPickList'
      summary: 商户提领代发详情
      tags:
      - 商户提领代发管理
  /merchantSend:
    post:
      parameters:
      - description: 商户代发
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.SendAdd'
      produces:
      - application/json
      responses:
        "200": {}
      summary: 代发
      tags:
      - 商户提领代发管理
  /notify:
    post:
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302": {}
      summary: 通知
      tags:
      - 商户提领代发管理
...
  /pick:
    get:
      parameters:
      - description: -1：等待中，1:执行中，2：成功，3：取消，4，失败
        in: query
        name: status
        type: integer
      - description: 0：所有，1：代发，2：提领
        in: query
        name: type
        type: integer
      - description: 商户id
        in: query
        name: userId
        type: integer
      - description: 起
        in: query
        name: from_date
        type: string
      - description: 至
        in: query
        name: to_date
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PickUpPage'
      summary: 管理端商户提领列表
      tags:
      - 商户提领代发管理
  /pick/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PickList'
      summary: 管理端商户提领详情
      tags:
      - 商户提领代发管理
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 提领订单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.SendEdit'
      produces:
      - application/json
      summary: 修改提领订单状态
      tags:
      - 商户提领代发管理
...
  /send:
    get:
      parameters:
      - description: -1：等待中，1:执行中，2：成功，3：取消，4，失败
        in: query
        name: status
        type: integer
      - description: 0：所有，1：代发，2：提领
        in: query
        name: type
        type: integer
      - description: 商户订单编号
        in: query
        name: orderCode
        type: string
      - description: 商户id
        in: query
        name: userId
        type: integer
      - description: 起
        in: query
        name: from_date
        type: string
      - description: 至
        in: query
        name: to_date
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页显示多少条
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SendPage'
      summary: 管理端商户代发列表
      tags:
      - 商户提领代发管理
  /send/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SendList'
      summary: 管理端商户代发详情
      tags:
      - 商户提领代发管理
...
  /system/{id}:
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 系统参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ParameterEdit'
      produces:
      - application/json
      summary: 更新系统参数
      tags:
      - 系统参数管理
  /systemGasPrice/{id}:
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 系统参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ParameterEdit'
      produces:
      - application/json
      summary: 更新GasPrice
      tags:
      - 系统参数管理
...
swagger: "2.0"